trabalho_final/central_server/requirements.txt
RPi.bme280==0.2.3
smbus2==0.4.0
RPi.GPIO==0.7.0
paho-mqtt==1.5.1
